+++ src/ring/middleware/format.clj
@@ -13,15 +13,16 @@
    use keywords as keys when deserializing. The first format is also
    the default serialization format (*:json* by default). You can also
    specify error-handlers for the params parsing with *:request-error-handler*
-   or the response encoding with *:response-error-handler*. See
-   [[ring.middleware.format-params/wrap-format-params]] and
+   or the response encoding with *:response-error-handler*.
+   Format options can be passed to responding middlewares using *:response-options*
+   and *:params-options*.
+   See [[ring.middleware.format-params/wrap-format-params]] and
    [[ring.middleware.format-response/wrap-format-response]] for details"
-  [handler & {:keys [formats response-error-handler request-error-handler]
+  [handler & {:keys [formats response-error-handler request-error-handler response-options params-options]
               :or {formats [:json :edn :msgpack :yaml :yaml-in-html
                             :transit-msgpack :transit-json]
                    response-error-handler res/default-handle-error
-                   request-error-handler par/default-handle-error}
-              :as options}]
+                   request-error-handler par/default-handle-error}}]
   (-> handler
-      (par/wrap-restful-params :formats formats :handle-error request-error-handler)
-      (res/wrap-restful-response :formats formats :handle-error response-error-handler)))
+      (par/wrap-restful-params :formats formats :handle-error request-error-handler :format-options params-options)
+      (res/wrap-restful-response :formats formats :handle-error response-error-handler :format-options response-options)))